llm-comfyui-captioning/requirements.txt
streamlit
transformers
torch
Pillow
googletrans==4.0.0-rc1
